+++ lib/src/core/common/navigation_pans/create_local_node_pane.dart
@@ -6,6 +6,7 @@ import 'package:gui/src/core/constants/app_constants.dart';
 import 'package:gui/src/core/enums/app_enums.dart';
 import 'package:gui/src/core/extensions/context_extensions.dart';
 import 'package:gui/src/core/utils/gen/localization/locale_keys.dart';
+import 'package:gui/src/data/models/fluent_navigation_state_model.dart';
 import 'package:gui/src/features/confirmation_seed/presentation/screen/confirmation_seed_screen.dart';
 import 'package:gui/src/features/finish/presentation/screen/finish_screen.dart';
 import 'package:gui/src/features/generation_seed/presentation/screens/generation_seed_screen.dart';
@@ -20,42 +21,18 @@ class CreateLocalNodePane extends StatelessWidget {
 
   @override
   Widget build(BuildContext context) {
-    return BlocBuilder<NavigationPaneCubit, int>(
-      builder: (context, selectedIndex) {
+    return BlocBuilder<NavigationPaneCubit, NavigationState>(
+      builder: (context, navigationState) {
         return AppLayout(
           content: NavigationView(
             pane: NavigationPane(
               displayMode: PaneDisplayMode.open,
               menuButton: const SizedBox(),
-              size: const NavigationPaneSize(openMaxWidth: 209),
-              selected: selectedIndex,
-              onChanged: (index) {
-                final stepValidationCubit = context.read<StepValidationCubit>();
-                final navigationCubit = context.read<NavigationPaneCubit>();
-
-                // Check if moving forward is allowed only if
-                // the current step is valid
-                final canGoForward = index == selectedIndex + 1 &&
-                    stepValidationCubit.isStepValid(selectedIndex);
-
-                // Allow moving backward only if you're not at the last page
-                final canGoBack = index == selectedIndex - 1 &&
-                    selectedIndex < AppConstants.createLocalNodeMaxIndex;
-
-                // If you've reached the last page, you won't be able to go back
-                if (selectedIndex == 5) {
-                  // If you've reached the last page, going backward
-                  // is not allowed
-                  if (index == selectedIndex - 1) {
-                    return;
-                  }
-                }
-
-                // Otherwise, allow moving forward or backward only if valid
-                if (canGoForward || canGoBack) {
-                  navigationCubit.setSelectedIndex(index);
-                }
-              },
+              size:
+                  const NavigationPaneSize(openMaxWidth: 209, compactWidth: 52),
+              selected: navigationState.selectedIndex,
+              onChanged: (index) =>
+                  _handleNavigationChange(context, navigationState, index),
               indicator: const SizedBox(),
               items: [
                 PaneItem(
@@ -64,7 +41,7 @@ class CreateLocalNodePane extends StatelessWidget {
                     context.tr(LocaleKeys.wallet_seed),
                     style: TextStyle(
                       color: context.detectPaneTextColor(
-                        isEnabledTextStyle: selectedIndex == 0,
+                        isEnabledTextStyle: navigationState.selectedIndex == 0,
                       ),
                     ),
                   ),
@@ -76,7 +53,7 @@ class CreateLocalNodePane extends StatelessWidget {
                     context.tr(LocaleKeys.confirm_seed),
                     style: TextStyle(
                       color: context.detectPaneTextColor(
-                        isEnabledTextStyle: selectedIndex == 1,
+                        isEnabledTextStyle: navigationState.selectedIndex == 1,
                       ),
                     ),
                   ),
@@ -88,7 +65,7 @@ class CreateLocalNodePane extends StatelessWidget {
                     context.tr(LocaleKeys.master_password),
                     style: TextStyle(
                       color: context.detectPaneTextColor(
-                        isEnabledTextStyle: selectedIndex == 2,
+                        isEnabledTextStyle: navigationState.selectedIndex == 2,
                       ),
                     ),
                   ),
@@ -100,7 +77,7 @@ class CreateLocalNodePane extends StatelessWidget {
                     context.tr(LocaleKeys.validator_config),
                     style: TextStyle(
                       color: context.detectPaneTextColor(
-                        isEnabledTextStyle: selectedIndex == 3,
+                        isEnabledTextStyle: navigationState.selectedIndex == 3,
                       ),
                     ),
                   ),
@@ -112,7 +89,7 @@ class CreateLocalNodePane extends StatelessWidget {
                     context.tr(LocaleKeys.initializing),
                     style: TextStyle(
                       color: context.detectPaneTextColor(
-                        isEnabledTextStyle: selectedIndex == 4,
+                        isEnabledTextStyle: navigationState.selectedIndex == 4,
                       ),
                     ),
                   ),
@@ -126,7 +103,7 @@ class CreateLocalNodePane extends StatelessWidget {
                     context.tr(LocaleKeys.finish),
                     style: TextStyle(
                       color: context.detectPaneTextColor(
-                        isEnabledTextStyle: selectedIndex == 5,
+                        isEnabledTextStyle: navigationState.selectedIndex == 5,
                       ),
                     ),
                   ),
@@ -138,5 +115,39 @@ class CreateLocalNodePane extends StatelessWidget {
         );
       },
     );
+  }
+
+  void _handleNavigationChange(
+    BuildContext context,
+    NavigationState navigationState,
+    int index,
+  ) {
+    final stepValidationCubit = context.read<StepValidationCubit>();
+    final navigationCubit = context.read<NavigationPaneCubit>();
+
+    // Check if moving forward is allowed only if
+    // the current step is valid
+    final canGoForward = index == navigationState.selectedIndex + 1 &&
+        stepValidationCubit.isStepValid(
+          navigationState.selectedIndex,
+        );
+
+    // Allow moving backward only if you're not at the last page
+    final canGoBack = index == navigationState.selectedIndex - 1 &&
+        navigationState.selectedIndex < AppConstants.createLocalNodeMaxIndex;
+
+    // If you've reached the last page, you won't be able to go back
+    if (navigationState.selectedIndex == 5) {
+      // If you've reached the last page, going backward
+      // is not allowed
+      if (index == navigationState.selectedIndex - 1) {
+        return;
+      }
+    }
+
+    // Otherwise, allow moving forward or backward only if valid
+    if (canGoForward || canGoBack) {
+      navigationCubit.setSelectedIndex(index);
+    }
   }
 }