+++ lib/src/features/main/navigation_pan_cubit/presentation/cubits/navigation_pan_cubit.dart
@@ -1,24 +1,8 @@
-import 'package:flutter_bloc/flutter_bloc.dart';
+import 'package:bloc/bloc.dart';
+import 'package:gui/src/data/models/fluent_navigation_state_model.dart';
 
-/// ## [NavigationPaneCubit] Class Documentation
-///
-/// The `NavigationPaneCubit` class manages the state of
-/// the selected index in a navigation pane. It extends `Cubit<int>`,
-/// where the integer value represents the current selected index.
-///
-/// ### Properties:
-///
-/// - **state**: The current selected index, which is an integer representing
-/// the active item in the navigation pane. The default value is 0 (first item).
-///
-/// ### Methods:
-///
-/// - **setSelectedIndex(int index)**: This method allows changing the selected
-/// index within the valid range (0 to 6). If the provided index is outside
-/// this range, the state will not change.
-
-class NavigationPaneCubit extends Cubit<int> {
-  NavigationPaneCubit() : super(0);
+class NavigationPaneCubit extends Cubit<NavigationState> {
+  NavigationPaneCubit() : super(NavigationState.initial());
   int _maxVisitedIndex = 0;
 
   void setSelectedIndex(int index) {
@@ -26,7 +10,21 @@ class NavigationPaneCubit extends Cubit<int> {
       if (index > _maxVisitedIndex) {
         _maxVisitedIndex = index;
       }
-      emit(index);
+      emit(
+        state.copyWith(selectedIndex: index),
+      );
     }
+  }
+
+  void toggleMenu() {
+    emit(state.copyWith(isMenuOpen: !state.isMenuOpen));
+  }
+
+  void openMenu() {
+    emit(state.copyWith(isMenuOpen: true));
+  }
+
+  void closeMenu() {
+    emit(state.copyWith(isMenuOpen: false));
   }
 }